requirements.txt
asgiref==3.9.1
cffi==1.17.1
cryptography==45.0.5
Django==5.2.4
djangorestframework==3.16.0
fernet==1.0.1
pyaes==1.6.1
pycparser==2.22
sqlparse==0.5.3
